Add new test_rmm_pytorch.py. Move fixtures to conftest.py.
